+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project are documented in this file.
    the default aspect ratio of the image.
  - deprecated `invoke_callback` functions in the slint interpreter in favor of `invoke`, they can also 
    invoke functions in addition to callback
+ - Propagate implicit constraints from child to parent, also when the child is not a layout (#783)
 
 ### Added